requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pdfplumber==0.11.4
pydantic==2.9.2
psycopg[binary]==3.2.3
numpy==2.1.2
google-generativeai==0.8.3
